match: +11 strict via fresh harvest wave (evaluation-order sculpting)
Regression-free wave from 4 Fable subagent lanes on main@11120, local-.cpp-only
(no header edits), whole-binary composed A/B run1==run2 = +11 / 0 regressions,
icf_alias_check HONEST (11 real-bodied):
- Str.cpp  String::operator=(char*) + operator+=(char*) -> 100 (manual copy loop
  with unsigned-int char temp reproduces retail's cmplwi vs strcpy's extsb)
- Sequence.cpp  RandomIntervalGroupSeqInst::Poll + ::ComputeNextTime -> 100
  (hoist *it into a local for unsigned null-check; hoist member loads in decl order)
- PostProc.cpp  RndPostProc::UpdateColorModulation -> 100 (opaque-ptr store blocks
  CSE of .x load across the .y store)
- GemTrackDir.cpp  GemTrackDir::GemPass -> 100 ((int) cast forces signed cmpwi)
- GuitarController.cpp  GuitarController::Handle -> 100 (per-TU HANDLE_MESSAGE
  redefine adds (unsigned char) on OnMsg return = retail bool truncation)
- ChordbookPanel.cpp  ChordbookPanel::CreateController -> 100 (lift nested sret temp
  into a named local)
- PrefabMgr.cpp  PrefabMgr::AssignPrefabsToSlots -> 100 (scale-into-index shape)
- BlockMgr.cpp  BlockMgr::Poll -> 100 (discard MainThread() result; explicit
  zero-init moves stat globals .bss->.data for retail anchor displacements)
- BandDirector.cpp  BandDirector::OnMidiShotCategory -> 100 (stmt-order + hoist
  PickDist temp before MakeString)

Dropped from the wave (coupled funclet regression, follow-up): MidiParser::AddMessage
(body hits 100 but the extra locals shift its EH-cleanup funclet's temp slot 0x50->0x54).
Killed (walls): SkinVertex (regalloc cycle), BinStream::Write (u8-narrow mask),
BlockMgr::AddTask (CSE reload). AppChild::Poll needs a Data.h Execute(bool)->Execute()
header change (DC3 drift) — deferred (shared-header, needs A/B).

+++ src/system/bandobj/GemTrackDir.cpp
@@ -692,7 +692,7 @@ void GemTrackDir::GemPass(int i1, int i2) {
     unk494++;
     if ((i2 & 1U) && mTrackInstrument == kInstDrum) {
         TrackInterface *par = BandTrack::mParent;
-        if (par) {
+        if ((int)par) {
             mKickPassCounter++;
             if (mKickPassCounter >= 8) {
                 float kickms = par->NextKickNoteMs();
